Mon, Oct 2, 2023, 12:02 PM +02:00

+++ filter-more/helpers.c
@@ -54,9 +54,9 @@ void blur(int height, int width, RGBTRIPLE image[height][width])
 
             if((i != 0) && (i != height-1) && (j != 0) && (j != width-1))
             {
-                int tot_red = 0;
-                int tot_green = 0;
-                int tot_blue = 0;
+                float tot_red = 0;
+                float tot_green = 0;
+                float tot_blue = 0;
 
                 for (int k = -1; k < 2; k++)
                 {
